+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 1.9.19 Release Candidate
 
+### Imrpovements
+
+- Use /data directory for temporary backup files instead of /tmp [#13853](https://github.com/opencrvs/opencrvs-core/issues/13853)
+- Add MinIO rsync backup option [#13851](https://github.com/opencrvs/opencrvs-core/issues/13851)
+
+
 ## 1.9.18
 
 ### Bug fixes